test: move application to fixtures
